Fix the realm test case
I inspected the signature base string in the original JS code. The base
string contained `alt=json` only once although the same parameter was
specified twice in a query string and in a body.

I think this is a bug in the original JS code. The original JS code
merges query and body parameters based on their keys. It can handle
duplicate keys within query or body, but it incorrectly overwrites body
parameters by query parameters if they share the same key. This behavior
is incorrect per 3.4.1.3.1. The tests succeeded because the test data
was tweaked to match the behavior of the implementation.

So, `alt=json` must not be included in the body in this test case.